+++ src/main/scala/subsystem/SystemBus.scala
@@ -4,30 +4,30 @@ package freechips.rocketchip.subsystem
 
 import Chisel._
 import freechips.rocketchip.config.{Parameters}
-import freechips.rocketchip.devices.tilelink.{DevNullParams, TLError, TLZero}
+import freechips.rocketchip.devices.tilelink._
 import freechips.rocketchip.diplomacy._
 import freechips.rocketchip.tilelink._
 import freechips.rocketchip.util._
 
 case class SystemBusParams(
-  beatBytes: Int,
-  blockBytes: Int,
-  policy: TLArbiter.Policy = TLArbiter.roundRobin,
-  errorDevice: Option[DevNullParams] = None) extends HasTLBusParams
+    beatBytes: Int,
+    blockBytes: Int,
+    policy: TLArbiter.Policy = TLArbiter.roundRobin,
+    zeroDevice: Option[AddressSet] = None,
+    errorDevice: Option[DevNullParams] = None)
+  extends HasTLBusParams with HasBuiltInDeviceParams
 
 class SystemBus(params: SystemBusParams)(implicit p: Parameters)
     extends TLBusWrapper(params, "system_bus")
+    with CanHaveBuiltInDevices
     with CanAttachTLSlaves
     with CanAttachTLMasters
     with HasTLXbarPhy {
+  attachBuiltInDevices(params)
 
   private val master_splitter = LazyModule(new TLSplitter)
   inwardNode :=* master_splitter.node
 
-  params.errorDevice.foreach { dnp => LazyScope("wrapped_error_device") {
-    val error = LazyModule(new TLError(params = dnp, beatBytes = params.beatBytes))
-    error.node := TLBuffer() := outwardNode
-  }}
   def busView = master_splitter.node.edges.in.head
 
   def toSplitSlave[D,U,E,B <: Data]
